Removing HTML help feature (for now just defanged) Adding "/nodefaultlib:libcmt" to debug configurations of WDS because of the luajit2 static lib Fixed the /debug:fastlink issue with VS2017

+++ windirstat/Controls/ownerdrawnlistcontrol.cpp
@@ -645,7 +645,7 @@ BOOL COwnerDrawnListControl::OnEraseBkgnd(CDC* pDC)
 
     CArray<int, int> columnOrder;
     columnOrder.SetSize(GetHeaderCtrl()->GetItemCount());
-    GetColumnOrderArray(columnOrder.GetData(), columnOrder.GetSize());
+    GetColumnOrderArray(columnOrder.GetData(), static_cast<int>(columnOrder.GetSize()));
 
     CArray<int, int> vertical;
     vertical.SetSize(GetHeaderCtrl()->GetItemCount());
